Feature/security

+++ app/Http/Controllers/ServicesController.php
@@ -16,8 +16,10 @@ class ServicesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($sal_id): JsonResponse
+    public function index(Request $request): JsonResponse
     {
+
+        $sal_id = $request->get('sal_id');
         $services = Services::where('sal_id', $sal_id)->get();
         return response()->json([
             'status' => 200,
@@ -35,14 +37,15 @@ class ServicesController extends Controller
      */
     public function create(Request $request)
     {
+
         try {
             DB::beginTransaction();
-           
-   
+                $sal_id = $request->get('sal_id');
                 $services = new Services;
-                $services = $services->create($request->all());
-          
-           
+                $services->sal_id = $sal_id;
+                $services = $services->create(array_merge($request->all(), ['sal_id' => $sal_id]));
+
+
             DB::commit();
             return response()->json([
                 'status' => 200,
@@ -73,26 +76,31 @@ class ServicesController extends Controller
     {
         try {
             DB::beginTransaction();
-            $services = Services::find($ser_id);
-            $services->update($request->all());
-            DB::commit();
-            return response()->json([
-                'status' => 200,
-                'message' => "Exitoso",
-                'services' => $services,
+            $sal_id = $request->get('sal_id');
+            if(Services::where('sal_id', $sal_id)->where('ser_id', $ser_id)->exists()){
+             $services = Services::where('sal_id', $sal_id)->where('ser_id', $ser_id)->first();
+             $services->update($request->all());
+             DB::commit();
+                 return response()->json([
+                        'status' => 200,
+                        'message' => "Exitoso",
+                        'services' => $services,
+                 ]);
+            }else{
+                 return response()->json([
+                        'status' => 400,
+                        'message' => "Error ninguna categoria",
+                 ],400);
+            }
 
-            ]);
         }catch (Exception $e){
             DB::rollBack();
             return response()->json([
                 'status' => 500,
                 'message' => "Error",
                 'error' => $e->getMessage(),
-
-            ]);
+            ],500);
         }
-
-
     }
 
     /**
@@ -103,19 +111,20 @@ class ServicesController extends Controller
     public function indexService($sto_id)
     {
         try {
-            $services = Services::all()->where('$sto_id',$sto_id);
+            $sal_id = $request->get('sal_id');
+            $services = Services::all()->where('sal_id',$sal_id);
             return response()->json([
                 'status' => 200,
                 'message' => "Exitoso",
                 'service' => $services,
-   
+
             ]);
         } catch (Exception $e) {
             return response()->json([
                 'status' => 500,
                 'message' => "Error",
                 'error' => $e->getMessage(),
-            
+
             ]);
         }
 
@@ -127,17 +136,26 @@ class ServicesController extends Controller
      * @param $ser_id
      * @return JsonResponse
      */
-    public function destroy($ser_id)
+    public function destroy(Request $request, $ser_id)
     {
        try {
             DB::beginTransaction();
-            $services = Services::find($ser_id);
-            $services->delete();
+             $sal_id = $request->get('sal_id');
+
+             if(Services::where('sal_id', $sal_id)->where('ser_id', $ser_id)->exists()){
+                $services = Services::where('sal_id', $sal_id)->where('ser_id', $ser_id)->first();
+                $services->delete();
+            }else{
+                return response()->json([
+                      'status' => 400,
+                      'message' => "No tienes este servicio",
+                      ]);
+            }
+
             DB::commit();
             return response()->json([
                 'status' => 200,
                 'message' => "Exitoso service delete",
-
             ]);
         }catch (Exception $e){
             DB::rollBack();
@@ -145,7 +163,6 @@ class ServicesController extends Controller
                 'status' => 500,
                 'message' => "Error",
                 'error' => $e->getMessage(),
-
             ]);
         }
     }
